% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_gen.R
\name{dat_gen}
\alias{dat_gen}
\title{Data generation function}
\usage{
dat_gen(N, V, es, seed = 1234)
}
\arguments{
\item{N}{sample size}

\item{V}{number of potential mediators}

\item{es}{effect size}

\item{seed}{seed. Default is 1234.}
}
\value{
dataset with 3 lists, containing the exposure X, the outcome Y and the potential V mediators M
}
\description{
This function generates the simulated data.
}
\details{
Title Data generation function
}
\examples{
dat_gen(100, 50, 1)
}
