% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC.R
\name{QUIC}
\alias{QUIC}
\title{Title QUIC from archive}
\usage{
QUIC(
  S,
  rho,
  path = NULL,
  tol = 1e-04,
  msg = 1,
  maxIter = 1000,
  X.init = NULL,
  W.init = NULL
)
}
\arguments{
\item{S}{S}

\item{rho}{rho}

\item{path}{path}

\item{tol}{tol}

\item{msg}{msg}

\item{maxIter}{maxIter}

\item{X.init}{X.init}

\item{W.init}{W.init}
}
\value{

}
\description{
Title QUIC from archive
}
\examples{
QUIC(Sigma2, rho = sigma2penalty*lam2[j], msg = 0)
}
