% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMInt-package.R
\docType{package}
\name{XMInt-package}
\alias{XMInt}
\alias{XMInt-package}
\title{XMInt: Model Selection for Multivariate Mediation with Interaction}
\description{
This package does model selection for multivariate mediation with interaction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ruiyangli1/XMInt}
  \item Report bugs at \url{https://github.com/ruiyangli1/XMInt/issues}
}

}
\author{
\strong{Maintainer}: Ruiyang Li \email{rl3034@cumc.columbia.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
