% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMInt_select.R
\name{XMInt_select}
\alias{XMInt_select}
\title{Model selection results}
\usage{
XMInt_select(
  X,
  Y,
  M,
  K = 20,
  zeta = 0.05,
  max.factor = 1.5,
  hbic_plot = FALSE,
  coef_print = FALSE
)
}
\arguments{
\item{X}{exposure}

\item{Y}{outcome}

\item{M}{mediators}

\item{K}{lambda sequence length (default = 20)}

\item{zeta}{min/max lambda ratio (default = 0.05)}

\item{max.factor}{enlarging factor (default = 1.5)}

\item{hbic_plot}{plot HBIC curve (default = \code{FALSE})}

\item{coef_print}{print coefficients (default = \code{FALSE})}
}
\value{
\emph{selected_mediator}: selected mediator

\emph{selected_interaction}: selected interaction

hbic_plot: HBIC plot for final model selection

hbic: computed HBIC

lambda: lambda sequence used

coefficient: estimated coefficients
}
\description{
This function gives the model selection results.
}
\examples{
data = dat_gen(200,100,es = 1)
X = data$X; Y = data$Y; M = data$M; I = X*M
XMInt_select(X,Y,M)
}
