% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_gen.R
\name{dat_gen}
\alias{dat_gen}
\title{Data generation function}
\usage{
dat_gen(N, V, es, seed = 1234)
}
\arguments{
\item{N}{sample size}

\item{V}{number of potential mediators}

\item{es}{effect size, representing the value of a, b1, b2 of the truth}

\item{seed}{seed. Default is 1234.}
}
\value{
The resulting dataset has 3 lists: X, Y and M
}
\description{
This function generates the simulation data (exposure X, outcome Y, and potential mediators M).

For each subject i = 1, ..., N:
\itemize{
\item \eqn{X_i \stackrel{i.i.d}{\sim} N(0, 1)}
\item \eqn{M_{i,v} = a_v X_i + e_{1_{i,v}}}, where \eqn{e_{1_{i,v}} \stackrel{i.i.d}{\sim} N(0, 1), v = 1, ..., V}
\item \eqn{Y_i = X_i + \sum_v b_{1_{v}} M_{i,v} + \sum_v b_{2_{v}} X_i \times M_{i,v} + e_{2_{i}}}, where \eqn{e_{2_{i}} \stackrel{i.i.d}{\sim} N(0, 1)}
}

The first three M variables ($M_1,M_2,M_3$) are set to be the true mediators (i.e., having non-zero $a$ and $b_1$ coefficients), $X \times M_1$ is set to be the true exposure-by-mediator interaction term (i.e., having non-zero $b_2$ coefficients), and all other coefficients are set to be 0.
}
\examples{
dat_gen(100, 50, 1)
}
