% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMInt-package.R
\docType{package}
\name{XMInt-package}
\alias{XMInt}
\alias{XMInt-package}
\title{XMInt: Model Selection for Exposure-Mediator Interaction}
\description{
This package contains functions for model selection for exposure-mediator interaction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ruiyangli1/XMInt}
  \item \url{https://ruiyangli1.github.io/XMInt}
  \item \url{https://ruiyangli1.github.io/XMInt/}
}

}
\keyword{internal}
