% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMInt-package.R
\docType{package}
\name{XMInt-package}
\alias{XMInt}
\alias{XMInt-package}
\title{XMInt: Model Selection for Multivariate Mediation with Interaction}
\description{
This package contains functions for model selection for multivariate mediation with interaction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ruiyangli1/XMInt}
  \item \url{https://ruiyangli1.github.io/XMInt}
}

}
\author{
\strong{Maintainer}: Ruiyang Li \email{rl3034@cumc.columbia.edu}

}
\keyword{internal}
